requirements.txt
asgiref==3.5.2
Django==4.1.3
psycopg2-binary==2.9.5
sqlparse==0.4.3
djangorestframework~=3.14.0